+++ changelog.md
@@ -7,6 +7,7 @@ Internal
 * Add test coverage for `client_query.py`.
 * Improve test coverage for `output.py`.
 * Add test coverage for `special/__init__.py`.
+* Add Apache Doris tests to `sqlexecute.py`.
 
 
 1.74.0 (2026/06/06)